+++ Lean.DataSource.DerivativeUniverseGenerator/AdditionalFieldGenerator.cs
@@ -13,96 +13,46 @@
  * limitations under the License.
 */
 
-using QuantConnect.Logging;
 using System;
-using System.Collections.Generic;
-using System.Globalization;
 using System.IO;
 using System.Linq;
 
 namespace QuantConnect.DataSource.DerivativeUniverseGenerator
 {
+    /// <summary>
+    /// Generate additional fields that needed to calculate from the whoel derivative chain
+    /// </summary>
     public class AdditionalFieldGenerator
     {
-        private const string impliedVolHeader = "impliedvolatility";
-        private const string deltaHeader = "delta";
-        private readonly DateTime _processingDate;
-        private readonly string _rootPath;
+        protected readonly DateTime _processingDate;
+        protected readonly string _rootPath;
 
+        /// <summary>
+        /// Instantiate a new instance of <see cref="AdditionalFieldGenerator"/>
+        /// </summary>
+        /// <param name="processingDate"></param>
+        /// <param name="rootPath"></param>
         public AdditionalFieldGenerator(DateTime processingDate, string rootPath)
         {
             _processingDate = processingDate;
             _rootPath = rootPath;
         }
 
-        public bool Run()
+        /// <summary>
+        /// Run the additional fields generation
+        /// </summary>
+        /// <returns>If the generator run successfully</returns>
+        public virtual bool Run()
         {
-            // per symbol
-            try
-            {
-                foreach (var subFolder in Directory.GetDirectories(_rootPath))
-                {
-                    var ivs = GetIvs(_processingDate, subFolder, out var latestFile);
-                    var additionalFields = new AdditionalFields(ivs);
-
-                    WriteToCsv(latestFile, additionalFields);
-                }
-            }
-            catch (Exception ex)
-            {
-                Log.Error(ex,
-                    $"AdditionalFieldGenerator.Run(): Error processing addtional fields for date {_processingDate:yyyy-MM-dd}");
-                return false;
-            }
-
-            return true;
+            throw new NotImplementedException("AdditionalFieldGenerator.Run(): Run method must be implemented.");
         }
 
-        private List<decimal> GetIvs(DateTime currentDateTime, string path, out string file)
-        {
-            // get i-year ATM IVs to calculate IV rank and percentile
-            var lastYearFiles = Directory.EnumerateFiles(path, "*.csv")
-                .AsParallel()
-                .Where(file => DateTime.TryParseExact(Path.GetFileNameWithoutExtension(file), "yyyyMMdd", null, DateTimeStyles.None, out var fileDate)
-                    && fileDate > currentDateTime.AddYears(-1)
-                    && fileDate <= currentDateTime)
-                .OrderBy(file => file)
-                .ToList();
-            file = lastYearFiles[^1];
-
-            return lastYearFiles.Select(csvFile => GetAtmIv(csvFile))
-                .ToList();
-        }
-
-        private decimal GetAtmIv(string csvPath)
-        {
-            var lines = File.ReadAllLines(csvPath)
-                .Where(s => !string.IsNullOrWhiteSpace(s))
-                .ToList();
-            var headers = lines[0].Split(',');
-            int deltaIndex = Array.IndexOf(headers, deltaHeader);
-            int ivIndex = Array.IndexOf(headers, impliedVolHeader);
-
-            if (deltaIndex == -1 || ivIndex == -1)
-            {
-                return 0m;
-            }
-
-            return Enumerable.Range(1, lines.Count - 1)
-                .AsParallel()
-                .Select(i =>
-                {
-                    var values = lines[i].Split(',');
-                    var delta = decimal.Parse(values[deltaIndex]);
-                    var iv = decimal.Parse(values[ivIndex]);
-                    return (Delta: delta, ImpliedVolatility: iv);
-                })
-                .OrderBy(x => Math.Abs(x.Delta - 0.5m))
-                .Select(x => x.ImpliedVolatility)
-                .First();
-        }
-
-        private void WriteToCsv(string csvPath, AdditionalFields additionalFields)
+        /// <summary>
+        /// Write the additional fields to the Csv file being generated
+        /// </summary>
+        /// <param name="csvPath">Target csv file path</param>
+        /// <param name="additionalFields">The addtional field content</param>
+        protected virtual void WriteToCsv(string csvPath, IAdditionalFields additionalFields)
         {
             var csv = File.ReadAllLines(csvPath)
                 .Where(s => !string.IsNullOrWhiteSpace(s))
@@ -111,11 +61,11 @@ namespace QuantConnect.DataSource.DerivativeUniverseGenerator
             {
                 if (i == 0)
                 {
-                    csv[i] += $",{AdditionalFields.Header}";
+                    csv[i] += $",{additionalFields.Header}";
                 }
                 else
                 {
-                    csv[i] += $",{additionalFields.IvRank},{additionalFields.IvPercentile}";
+                    csv[i] += $",{additionalFields.Content}";
                 }
             }
