+++ Lean.DataSource.DerivativeUniverseGenerator/IAdditionalFields.cs
@@ -1,0 +1,33 @@
+﻿/*
+ * QUANTCONNECT.COM - Democratizing Finance, Empowering Individuals.
+ * Lean Algorithmic Trading Engine v2.0. Copyright 2024 QuantConnect Corporation.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+*/
+
+namespace QuantConnect.DataSource.DerivativeUniverseGenerator
+{
+    /// <summary>
+    /// Interface of additional fields which requires the whole derivative chain to calculate from
+    /// </summary>
+    public interface IAdditionalFields
+    {
+        /// <summary>
+        /// Csv header
+        /// </summary>
+        string Header { get; }
+
+        /// <summary>
+        /// Fields calculated, in form of comma-separated string
+        /// </summary>
+        string Content { get; }
+    }
+}
