+++ Lean.DataSource.OptionsUniverseGenerator/Program.cs
@@ -39,5 +39,10 @@ namespace QuantConnect.DataSource.OptionsUniverseGenerator
         {
             return new OptionsUniverseGenerator(processingDate, securityType, market, dataFolderRoot, outputFolderRoot);
         }
+
+        protected override DerivativeUniverseGenerator.AdditionalFieldGenerator GetAdditionalFieldGenerator(DateTime processingDate, string rootFolder)
+        {
+            return new OptionAdditionalFieldGenerator(processingDate, rootFolder);
+        }
     }
 }
